+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* 3/2/2016, [#5](https://github.com/dblock/slack-shellbot/issues/5): `:q` will quit `vi` without saving the file - [@dblock](https://github.com/dblock).
 * 3/2/2016, [#4](https://github.com/dblock/slack-shellbot/issues/4): Edit messages in-place with `vi` - [@dblock](https://github.com/dblock).
 * 2/11/2016: Use OOB GC, swapped Puma for Unicorn - [@dblock](https://github.com/dblock).
 * 1/7/2016: Added support for quoted file names - [@dblock](https://github.com/dblock).